+++ src/main/java/com/github/dockerjava/api/model/Container.java
@@ -5,14 +5,12 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.github.dockerjava.api.command.ListContainersCmd;
-import com.github.dockerjava.api.model.Network.Ipam;
 import com.github.dockerjava.core.RemoteApiVersion;
 import org.apache.commons.lang.builder.EqualsBuilder;
 import org.apache.commons.lang.builder.HashCodeBuilder;
 import org.apache.commons.lang.builder.ToStringBuilder;
 
 import javax.annotation.CheckForNull;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -46,7 +44,7 @@ public class Container {
     private String[] names;
 
     @JsonProperty("Ports")
-    public Port[] ports;
+    public ContainerPort[] ports;
 
     @JsonProperty("Labels")
     public Map<String, String> labels;
@@ -72,7 +70,7 @@ public class Container {
      * @since ~{@link RemoteApiVersion#VERSION_1_20}
      */
     @JsonProperty("HostConfig")
-    private HostConfig hostConfig;
+    private ContainerHostConfig hostConfig;
 
     /**
      * Docker API docs says "list of networks", but json names `networkSettings`.
@@ -81,7 +79,7 @@ public class Container {
      * @since ~{@link RemoteApiVersion#VERSION_1_22}
      */
     @JsonProperty("NetworkSettings")
-    private NetworkSettings networkSettings;
+    private ContainerNetworkSettings networkSettings;
 
     public String getId() {
         return id;
@@ -108,7 +106,7 @@ public class Container {
         return status;
     }
 
-    public Port[] getPorts() {
+    public ContainerPort[] getPorts() {
         return ports;
     }
 
@@ -140,7 +138,7 @@ public class Container {
      * @see #networkSettings
      */
     @CheckForNull
-    public NetworkSettings getNetworkSettings() {
+    public ContainerNetworkSettings getNetworkSettings() {
         return networkSettings;
     }
 
@@ -148,7 +146,7 @@ public class Container {
      * @see #hostConfig
      */
     @CheckForNull
-    public HostConfig getHostConfig() {
+    public ContainerHostConfig getHostConfig() {
         return hostConfig;
     }
 
@@ -199,297 +197,5 @@ public class Container {
                 .append(hostConfig)
                 .append(networkSettings)
                 .toHashCode();
-    }
-
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    public static final class HostConfig {
-        @JsonProperty("NetworkMode")
-        private String networkMode;
-
-        public String getNetworkMode() {
-            return networkMode;
-        }
-    }
-
-    /**
-     * @since {@link RemoteApiVersion#VERSION_1_22}
-     */
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    public static final class NetworkSettings {
-        @JsonProperty("Networks")
-        private Map<String, Network> networks;
-
-        public Map<String, Network> getNetworks() {
-            return networks;
-        }
-
-        @JsonIgnoreProperties(ignoreUnknown = true)
-        public static final class Network {
-            /**
-             * FIXME verify
-             */
-            @JsonProperty("IPAMConfig")
-            private Ipam.Config ipamConfig;
-
-            /**
-             * FIXME verify
-             */
-            @JsonProperty("Links")
-            private List<String> links;
-
-            /**
-             * FIXME no docs, unknown field.
-             * Type picked from `docker/vendor/src/github.com/docker/engine-api/types/network/network.go`
-             * @since {@link RemoteApiVersion#VERSION_1_22}
-             */
-            @JsonProperty("Aliases")
-            private List<String> aliases;
-
-            @JsonProperty("NetworkID")
-            private String networkID;
-
-            @JsonProperty("EndpointID")
-            private String endpointId;
-
-            @JsonProperty("Gateway")
-            private String gateway;
-
-            @JsonProperty("IPAddress")
-            private String ipAddress;
-
-            @JsonProperty("IPPrefixLen")
-            private Integer ipPrefixLen;
-
-            @JsonProperty("IPv6Gateway")
-            private String ipV6Gateway;
-
-            @JsonProperty("GlobalIPv6Address")
-            private String globalIPv6Address;
-
-            @JsonProperty("GlobalIPv6PrefixLen")
-            private Integer globalIPv6PrefixLen;
-
-            @JsonProperty("MacAddress")
-            private String macAddress;
-
-            /**
-             * @see #aliases
-             */
-            @CheckForNull
-            public List<String> getAliases() {
-                return aliases;
-            }
-
-            /**
-             * @see #endpointId
-             */
-            @CheckForNull
-            public String getEndpointId() {
-                return endpointId;
-            }
-
-            /**
-             * @see #gateway
-             */
-            @CheckForNull
-            public String getGateway() {
-                return gateway;
-            }
-
-            /**
-             * @see #globalIPv6Address
-             */
-            @CheckForNull
-            public String getGlobalIPv6Address() {
-                return globalIPv6Address;
-            }
-
-            /**
-             * @see #globalIPv6PrefixLen
-             */
-            @CheckForNull
-            public Integer getGlobalIPv6PrefixLen() {
-                return globalIPv6PrefixLen;
-            }
-
-            /**
-             * @see #ipAddress
-             */
-            @CheckForNull
-            public String getIpAddress() {
-                return ipAddress;
-            }
-
-            /**
-             * @see #ipamConfig
-             */
-            @CheckForNull
-            public Ipam.Config getIpamConfig() {
-                return ipamConfig;
-            }
-
-            /**
-             * @see #ipPrefixLen
-             */
-            @CheckForNull
-            public Integer getIpPrefixLen() {
-                return ipPrefixLen;
-            }
-
-            /**
-             * @see #ipV6Gateway
-             */
-            @CheckForNull
-            public String getIpV6Gateway() {
-                return ipV6Gateway;
-            }
-
-            /**
-             * @see #links
-             */
-            @CheckForNull
-            public List<String> getLinks() {
-                return links;
-            }
-
-            /**
-             * @see #macAddress
-             */
-            @CheckForNull
-            public String getMacAddress() {
-                return macAddress;
-            }
-
-            /**
-             * @see #networkID
-             */
-            @CheckForNull
-            public String getNetworkID() {
-                return networkID;
-            }
-
-            @Override
-            public String toString() {
-                return new ToStringBuilder(this)
-                        .append("aliases", aliases)
-                        .append("ipamConfig", ipamConfig)
-                        .append("links", links)
-                        .append("networkID", networkID)
-                        .append("endpointId", endpointId)
-                        .append("gateway", gateway)
-                        .append("ipAddress", ipAddress)
-                        .append("ipPrefixLen", ipPrefixLen)
-                        .append("ipV6Gateway", ipV6Gateway)
-                        .append("globalIPv6Address", globalIPv6Address)
-                        .append("globalIPv6PrefixLen", globalIPv6PrefixLen)
-                        .append("macAddress", macAddress)
-                        .toString();
-            }
-
-            @Override
-            public boolean equals(Object o) {
-                if (this == o) return true;
-
-                if (o == null || getClass() != o.getClass()) return false;
-
-                Network network = (Network) o;
-
-                return new EqualsBuilder()
-                        .append(ipamConfig, network.ipamConfig)
-                        .append(links, network.links)
-                        .append(aliases, network.aliases)
-                        .append(networkID, network.networkID)
-                        .append(endpointId, network.endpointId)
-                        .append(gateway, network.gateway)
-                        .append(ipAddress, network.ipAddress)
-                        .append(ipPrefixLen, network.ipPrefixLen)
-                        .append(ipV6Gateway, network.ipV6Gateway)
-                        .append(globalIPv6Address, network.globalIPv6Address)
-                        .append(globalIPv6PrefixLen, network.globalIPv6PrefixLen)
-                        .append(macAddress, network.macAddress)
-                        .isEquals();
-            }
-
-            @Override
-            public int hashCode() {
-                return new HashCodeBuilder(17, 37)
-                        .append(ipamConfig)
-                        .append(links)
-                        .append(aliases)
-                        .append(networkID)
-                        .append(endpointId)
-                        .append(gateway)
-                        .append(ipAddress)
-                        .append(ipPrefixLen)
-                        .append(ipV6Gateway)
-                        .append(globalIPv6Address)
-                        .append(globalIPv6PrefixLen)
-                        .append(macAddress)
-                        .toHashCode();
-            }
-        }
-    }
-
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    public static class Port {
-
-        @JsonProperty("IP")
-        private String ip;
-
-        @JsonProperty("PrivatePort")
-        private Integer privatePort;
-
-        @JsonProperty("PublicPort")
-        private Integer publicPort;
-
-        @JsonProperty("Type")
-        private String type;
-
-        public String getIp() {
-            return ip;
-        }
-
-        public Integer getPrivatePort() {
-            return privatePort;
-        }
-
-        public Integer getPublicPort() {
-            return publicPort;
-        }
-
-        public String getType() {
-            return type;
-        }
-
-        @Override
-        public String toString() {
-            return ToStringBuilder.reflectionToString(this);
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-
-            if (o == null || getClass() != o.getClass()) return false;
-
-            Port port = (Port) o;
-
-            return new EqualsBuilder()
-                    .append(ip, port.ip)
-                    .append(privatePort, port.privatePort)
-                    .append(publicPort, port.publicPort)
-                    .append(type, port.type)
-                    .isEquals();
-        }
-
-        @Override
-        public int hashCode() {
-            return new HashCodeBuilder(17, 37)
-                    .append(ip)
-                    .append(privatePort)
-                    .append(publicPort)
-                    .append(type)
-                    .toHashCode();
-        }
     }
 }
